requirements.txt
ccxt
pandas
pandas-ta
psycopg2-binary
python-telegram-bot
python-dotenv
python-dotenv

